move ThemeValues to `hooks`
