+++ packages/app-staking/src/canvas/Pool/Overview/Performance/InactiveGraph.tsx
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { getStakingChainData } from 'consts/util'
-import { useThemeValues } from 'contexts/ThemeValues'
 import { useDateFormat } from 'hooks/useDateFormat'
 import { useNetwork } from 'hooks/useNetwork'
+import { useThemeValues } from 'hooks/useThemeValues'
 import { useTranslation } from 'react-i18next'
 import { PayoutLine } from 'ui-graphs'
