+++ packages/app-staking/src/canvas/ValidatorMetrics/EraPoints/ActiveGraph.tsx
@@ -1,8 +1,8 @@
 // Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { useThemeValues } from 'contexts/ThemeValues'
 import { useDateFormat } from 'hooks/useDateFormat'
+import { useThemeValues } from 'hooks/useThemeValues'
 import { useValidatorEraPoints } from 'plugin-staking-api'
 import { useTranslation } from 'react-i18next'
 import type { NetworkId } from 'types'
