+++ packages/app-staking/src/canvas/ValidatorMetrics/ValidatorGeo.tsx
@@ -5,9 +5,9 @@ import { PolkawatchApi, type ValidatorDetail } from '@polkawatch/ddp-client'
 import { useSize } from '@w3ux/hooks'
 import { PolkawatchConfig } from 'consts/plugins'
 import { getPolkawatchConfig } from 'consts/util'
-import { useThemeValues } from 'contexts/ThemeValues'
 import { useNetwork } from 'hooks/useNetwork'
 import { usePlugins } from 'hooks/usePlugins'
+import { useThemeValues } from 'hooks/useThemeValues'
 import { useUi } from 'hooks/useUi'
 import { CardWrapper } from 'library/Card/Wrappers'
 import { StatusLabel } from 'library/StatusLabel'
