+++ packages/app-staking/src/library/ValidatorList/index.tsx
@@ -7,13 +7,13 @@ import { useActiveAccount } from '@polkadot-cloud/connect'
 import { getPeopleChainId } from 'consts/util'
 import { useFilters } from 'contexts/Filters'
 import { ListProvider, useList } from 'contexts/List'
-import { useThemeValues } from 'contexts/ThemeValues'
 import type { ValidatorListEntry } from 'contexts/Validators/types'
 import { useValidators } from 'contexts/Validators/ValidatorEntries'
 import { useApi } from 'hooks/useApi'
 import { useNetwork } from 'hooks/useNetwork'
 import { usePlugins } from 'hooks/usePlugins'
 import { useSyncing } from 'hooks/useSyncing'
+import { useThemeValues } from 'hooks/useThemeValues'
 import { useValidatorRewardRateBatch } from 'hooks/useValidatorRewardRateBatch'
 import { FilterHeaderWrapper, List, Wrapper as ListWrapper } from 'library/List'
 import { MotionContainer } from 'library/List/MotionContainer'
