+++ packages/app-staking/src/pages/Nominate/NominationGeo/index.tsx
@@ -5,11 +5,11 @@ import { useActiveAccount } from '@polkadot-cloud/connect'
 import { PolkawatchApi } from '@polkawatch/ddp-client'
 import { PolkawatchConfig } from 'consts/plugins'
 import { getPolkawatchConfig } from 'consts/util'
-import { useThemeValues } from 'contexts/ThemeValues'
 import { useHelp } from 'hooks/useHelp'
 import { useNetwork } from 'hooks/useNetwork'
 import { usePlugins } from 'hooks/usePlugins'
 import { useStaking } from 'hooks/useStaking'
+import { useThemeValues } from 'hooks/useThemeValues'
 import { CardWrapper } from 'library/Card/Wrappers'
 import { StatusLabel } from 'library/StatusLabel'
 import { useEffect, useState } from 'react'
