+++ packages/app-staking/src/pages/Pools/PoolShares/index.tsx
@@ -8,7 +8,6 @@ import { getChainIcons } from 'assets'
 import BigNumber from 'bignumber.js'
 import { PoolSharesDays } from 'consts'
 import { getStakingChainData, isPoolShareEnabled } from 'consts/util'
-import { useThemeValues } from 'contexts/ThemeValues'
 import { getUnixTime } from 'date-fns'
 import { useActivePool } from 'hooks/useActivePool'
 import { useCurrency } from 'hooks/useCurrency'
@@ -16,6 +15,7 @@ import { useDateFormat } from 'hooks/useDateFormat'
 import { useNetwork } from 'hooks/useNetwork'
 import { usePlugins } from 'hooks/usePlugins'
 import { useSyncing } from 'hooks/useSyncing'
+import { useThemeValues } from 'hooks/useThemeValues'
 import { Balance } from 'library/Balance'
 import { CardWrapper } from 'library/Card/Wrappers'
 import { StatusLabel } from 'library/StatusLabel'
