+++ packages/app-staking/src/pages/Rewards/PayoutList/Inner.tsx
@@ -8,12 +8,12 @@ import BigNumber from 'bignumber.js'
 import { getStakingChainData } from 'consts/util'
 import { ListProvider, useList } from 'contexts/List'
 import { useBondedPools } from 'contexts/Pools/BondedPools'
-import { useThemeValues } from 'contexts/ThemeValues'
 import { useValidators } from 'contexts/Validators/ValidatorEntries'
 import { formatDistance, fromUnixTime } from 'date-fns'
 import { useApi } from 'hooks/useApi'
 import { useDateFormat } from 'hooks/useDateFormat'
 import { useNetwork } from 'hooks/useNetwork'
+import { useThemeValues } from 'hooks/useThemeValues'
 import { Header, List, Wrapper as ListWrapper } from 'library/List'
 import { MotionContainer } from 'library/List/MotionContainer'
 import { Pagination } from 'library/List/Pagination'
