+++ packages/hooks/src/index.ts
@@ -32,6 +32,7 @@ export * from './useStakingMetrics'
 export * from './useSupplyStaked'
 export * from './useSyncing'
 export * from './useTheme'
+export * from './useThemeValues'
 export * from './useTokenPrices'
 export * from './useTooltip'
 export * from './useTxMeta'
