+++ packages/hooks/src/useThemeValues/types.ts
@@ -1,6 +1,6 @@
 // Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-export interface ThemeValuesContextInterface {
+export interface ThemeValuesHookInterface {
 	getThemeValue: (name: string) => string
 }
